// src/app/_services/model/Attendance/PermissionConfiguration.ts
import { BaseModel } from "../Common/BaseModel";


export class PermissionConfiguration extends BaseModel {

    CompanyId: any;
    ClientId: any;
    ClientContractId: any;
    TeamId: any;
    EmployeeId: any;
    IsAllowedToRequestPermission: any;
    IsAllowedToSubmitPostMaxLimit: any;
    MaxNoOfPastDaysAllowed: any;
    MinumunTimeIntervalBetweenPermission: any;
    Status: any;
    MinTimeInterval: any
    MaxTimeInterval: any


}